PLX-1005 api refactored for board and calendar naming, fixed api controller structure from PLX-1003
