+++ src/apiControllers/__tests__/getBoard.test.ts
@@ -1,16 +1,17 @@
-import { getByUserProjectId /* getByUserId, createItem */ } from '../index'
+import { getBoard } from '../getBoard'
 
-import project1001FullBoardApi from '../__data__/project-1001-full-board-api.json'
+import project1001FullBoardApi from '../__data__/project-1001-full-api.json'
 
-import * as PlaaxItemsRepo from '../../../repositories/PlaaxItemsRepo'
-import { jsonToProjectFull } from '../test.utils'
+import * as PlaaxItemsRepo from '../../repositories/PlaaxItemsRepo'
+import { jsonToProjectFull } from './test.utils'
 
-jest.mock('../../../repositories/PlaaxItemsRepo')
+jest.mock('../../repositories/PlaaxItemsRepo')
 const mockedPlaaxItemsRepo = PlaaxItemsRepo as jest.Mocked<typeof PlaaxItemsRepo>
 
 describe('API Controller: board', () => {
   describe('GET user board of 1 project', () => {
     it('should respond with 200', async () => {
+      // todo: remove ctx because is not infra-detached (related to Koa)
       const ctx = {
         params: { userId: '1005', projectId: '1001' },
         status: undefined,
@@ -23,11 +24,13 @@ describe('API Controller: board', () => {
         return jsonToProjectFull(project1001FullBoardApi)
       })
 
-      await getByUserProjectId(ctx)
+      await getBoard('1234')
 
       expect(mockedPlaaxItemsRepo.getUserProjectBoard).toHaveBeenCalledTimes(1) // todo: called with...
       expect(ctx.body).toMatchObject(project1001FullBoardApi)
       expect(ctx.status).toEqual(200)
     })
   })
+
+  // describe('GET full user board with many projects', () => {})
 })
