--- src/apiControllers/board/__tests__/index-pre.test.ts.notes
@@ -1,35 +1,0 @@
-// import { getByUserId, getByUserProjectId, createItem } from '../index'
-//
-// import * as PlaaxItemsRepo from '../../../repositories/PlaaxItemsRepo'
-//
-// jest.mock('../../../repositories/PlaaxItemsRepo')
-//
-// describe('API Controller: board', () => {
-//   describe('GET user board of 1 project', () => {
-//     it('should respond with 200', async () => {
-//       const ctx = { params: { userId: '1005', projectId: '1001' }, status: undefined }
-//       await getByUserProjectId(ctx)
-//       expect(PlaaxItemsRepo.getUserProjectBoard).toHaveBeenCalledTimes(1)
-//       // expect(ctx.body).toMatchObject()
-//       expect(ctx.status).toEqual(200)
-//     })
-//   })
-//
-//   describe('GET full user board with many projects', () => {
-//     it('should respond with 200', async () => {
-//       const ctx = { params: { userId: '1005' }, status: undefined }
-//       await getByUserId(ctx)
-//       expect(PlaaxItemsRepo.getFullUserBoard).toHaveBeenCalledTimes(1)
-//       // expect(ctx.body).toMatchObject()
-//       expect(ctx.status).toEqual(200)
-//     })
-//   })
-//
-//   describe('POST an item to user board for 1 project', () => {
-//     it('should respond with 200', async () => {
-//       // const ctx = { params: { id: '11-22-33' }, status: undefined }
-//       // await createItem(ctx)
-//       // expect(ctx.status).toEqual(200)
-//     })
-//   })
-// })
