--- src/apiControllers/calendar/index.ts
@@ -1,14 +1,0 @@
-export const getByUserId = async (ctx): Promise<void> => {
-  ctx.status = 200
-  ctx.body = { }
-}
-
-export const getByUserProjectId = async (ctx): Promise<void> => {
-  ctx.status = 200
-  ctx.body = { }
-}
-
-export const addItemToCalendar = async (ctx): Promise<void> => {
-  ctx.status = 200
-  ctx.body = { }
-}
