+++ src/apiControllers/getBoard.ts
@@ -1,0 +1,4 @@
+export const getBoard = async (projectId?: string): Promise<number> => {
+  // if !projectId or !validation => throw error!
+  return 1
+}
