--- src/apiControllers/getFullProject.ts
@@ -1,4 +1,0 @@
-export const getFullProject = async (projectId?: string): Promise<number> => {
-  // if !projectId or !validation => throw error!
-  return 1
-}
