+++ src/apiControllers/getUserProjectsList.ts
@@ -1,0 +1,4 @@
+export const getUserProjectsList = async (projectId?: string): Promise<number> => {
+  // if !projectId or !validation => throw error!
+  return 1
+}
